the owner only than can delete the lobby

+++ catan/lobby/views.py
@@ -137,5 +137,11 @@ class RoomsView(viewsets.ModelViewSet):
             )
 
         room = self.get_object()
+        if request.user != room.owner:
+            return Response(
+                'The owner is the only than can delete this room',
+                status=status.HTTP_406_NOT_ACCEPTABLE
+            )
+
         room.delete()
         return Response()
